+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 4.1 FIDASIM weight
+- Area normalization not applied, int FIDASIM markers, as the pinhole size is different in FIDASIM than in SINPA to speed up the simulation
 ## 4.0 Torwards interfaces
 - First step to move torwards interfaces and forget about dummy ifs
 - New strike file formats added